fix: removed healthcheck as wget/curl are not available in the distroless image
This is done for security reasons as the availability of wget/curl may allow an
attacker to download additional tooling.

This should not be a problem when using advanced container orchestrators
like Kubernetes. A workaround for other execution environments would be
the introduction of a sidecar which checks the /health endpoint of the
gateway container, reporting it as its own health status.